backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
httpx>=0.27.0
python-multipart>=0.0.9
Pillow>=10.3.0
pytesseract>=0.3.13
python-dotenv>=1.0.0
pydantic>=2.7.0
